Python Script for Automation
This Python Script is for shack automation. The script polls my K3 every 2 seconds and looks for band changes. If the band changes, the WR9R Antenna Switch setting is retrieved from the lookup table and the Switch is changed, over the network, to match the newly selected band on the K3.